fix(eggs): messages with reactions don't get deleted

+++ attubot/tasks/egg_cleanup.py
@@ -67,6 +67,8 @@ class EggCleanupTask(BaseTask):
                     break
                 if message.id in egg_message_ids:
                     continue
+                if message.reactions:
+                    continue
                 try:
                     await message.delete()
                     total_deleted += 1
